Refactor index.js to load modules + plugins
