+++ LyraCore/db.json
@@ -1,0 +1,12 @@
+{
+  "levels": {
+    "1234": {
+      "totalXp": 0,
+      "guilds": {}
+    },
+    "12323232": {
+      "totalXp": 0,
+      "guilds": {}
+    }
+  }
+}
