--- LyraCore/index.js
@@ -1,9 +1,0 @@
-const logger = require("./logging");
-const baseEmbed = require("./BaseEmbed");
-const buildCommand = require("./buildCommand");
-
-module.exports = {
-    logger,
-    baseEmbed,
-    buildCommand,
-}
