+++ LyraCore/plugins.js
@@ -1,0 +1,8 @@
+const Path = require("path");
+
+const plugins = {};
+console.log(Path.resolve(__dirname, "."))
+
+module.exports = function loadPlugins(pluginApi) {
+    return {};
+}
