+++ plugins/status.js
@@ -1,0 +1,11 @@
+module.exports = function ({CommandoClient}) {
+    // client.once('ready', () => {
+    //     client.user.setActivity('intently | use ly!ra', {type: 'WATCHING'})
+    //     setInterval(() => {
+    //         client.user.setActivity('intently | use ly!ra', {type: 'WATCHING'})
+    //     }, 10000);
+    //     console.log(`${client.user.tag} is Ready!`);
+    //     const Guilds = client.guilds.cache.map(guild => guild.name);
+    //     console.log(Guilds, 'Connected!');
+    // });
+}
